Baseline/requirements.txt
speedtest-cli
pandas
plotly
streamlit
